Reorganise how timer strings to be saved in HDF5 are created (#200)
Julia internally changed how `rpad` computes the length of the string, which
breaks the logic we used here.  With this change we sidestep this issue by
gathering a vector of the actual lengths of all timers in all ranks, instead of
padding all of them.  This is also more accurate and avoids writing unnecessary
spaces at the end of the strings.